Move CalendarFields to individual module + related cleanup
