+++ src/builtins/core/calendar/types.rs
@@ -1,252 +1,16 @@
 //! Implementation of `ResolvedCalendarFields`
 
-use ixdtf::records::DateRecord;
 use tinystr::tinystr;
 use tinystr::TinyAsciiStr;
 
 use alloc::format;
 
+use crate::fields::CalendarFields;
 use crate::iso::{constrain_iso_day, is_valid_iso_day};
 use crate::options::ArithmeticOverflow;
-use crate::PlainMonthDay;
-use crate::PlainYearMonth;
-use crate::{Calendar, PlainDate, PlainDateTime};
+use crate::Calendar;
 use crate::{TemporalError, TemporalResult};
 use icu_calendar::AnyCalendarKind;
-
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct CalendarFields {
-    // A potentially set `year` field.
-    pub year: Option<i32>,
-    // A potentially set `month` field.
-    pub month: Option<u8>,
-    // A potentially set `month_code` field.
-    pub month_code: Option<MonthCode>,
-    // A potentially set `day` field.
-    pub day: Option<u8>,
-    // A potentially set `era` field.
-    pub era: Option<TinyAsciiStr<19>>,
-    // A potentially set `era_year` field.
-    pub era_year: Option<i32>,
-}
-
-impl CalendarFields {
-    pub const fn new() -> Self {
-        Self {
-            year: None,
-            month: None,
-            month_code: None,
-            day: None,
-            era: None,
-            era_year: None,
-        }
-    }
-
-    pub const fn with_era(mut self, era: Option<TinyAsciiStr<19>>) -> Self {
-        self.era = era;
-        self
-    }
-
-    pub const fn with_era_year(mut self, era_year: Option<i32>) -> Self {
-        self.era_year = era_year;
-        self
-    }
-
-    pub const fn with_year(mut self, year: i32) -> Self {
-        self.year = Some(year);
-        self
-    }
-
-    pub const fn with_optional_year(mut self, year: Option<i32>) -> Self {
-        self.year = year;
-        self
-    }
-
-    pub const fn with_month(mut self, month: u8) -> Self {
-        self.month = Some(month);
-        self
-    }
-
-    pub const fn with_optional_month(mut self, month: Option<u8>) -> Self {
-        self.month = month;
-        self
-    }
-
-    pub const fn with_month_code(mut self, month_code: MonthCode) -> Self {
-        self.month_code = Some(month_code);
-        self
-    }
-
-    pub const fn with_optional_month_code(mut self, month_code: Option<MonthCode>) -> Self {
-        self.month_code = month_code;
-        self
-    }
-
-    pub const fn with_day(mut self, day: u8) -> Self {
-        self.day = Some(day);
-        self
-    }
-
-    pub const fn with_optional_day(mut self, day: Option<u8>) -> Self {
-        self.day = day;
-        self
-    }
-}
-
-impl CalendarFields {
-    pub fn is_empty(&self) -> bool {
-        *self == Self::new()
-    }
-
-    pub(crate) fn from_date_record(date_record: DateRecord) -> Self {
-        Self {
-            year: Some(date_record.year),
-            month: Some(date_record.month),
-            month_code: None,
-            day: Some(date_record.day),
-            era: None,
-            era_year: None,
-        }
-    }
-
-    pub(crate) fn from_month_day(month_day: &PlainMonthDay) -> Self {
-        Self {
-            year: None,
-            month: None,
-            month_code: Some(month_day.month_code()),
-            era: None,
-            era_year: None,
-            day: Some(month_day.day()),
-        }
-    }
-
-    crate::impl_with_fallback_method!(with_fallback_date, CalendarFields, (with_day: day) PlainDate);
-    crate::impl_with_fallback_method!(with_fallback_datetime, CalendarFields, (with_day:day) PlainDateTime);
-    crate::impl_field_keys_to_ignore!((with_day:day));
-}
-
-impl From<&YearMonthCalendarFields> for CalendarFields {
-    fn from(value: &YearMonthCalendarFields) -> Self {
-        Self {
-            year: value.year,
-            month: value.month,
-            month_code: value.month_code,
-            day: None,
-            era: value.era,
-            era_year: value.era_year,
-        }
-    }
-}
-
-impl From<&CalendarFields> for YearMonthCalendarFields {
-    fn from(value: &CalendarFields) -> Self {
-        Self {
-            year: value.year,
-            month: value.month,
-            month_code: value.month_code,
-            era: value.era,
-            era_year: value.era_year,
-        }
-    }
-}
-
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct YearMonthCalendarFields {
-    // A potentially set `year` field.
-    pub year: Option<i32>,
-    // A potentially set `month` field.
-    pub month: Option<u8>,
-    // A potentially set `month_code` field.
-    pub month_code: Option<MonthCode>,
-    // A potentially set `era` field.
-    pub era: Option<TinyAsciiStr<19>>,
-    // A potentially set `era_year` field.
-    pub era_year: Option<i32>,
-}
-
-impl YearMonthCalendarFields {
-    pub const fn new() -> Self {
-        Self {
-            year: None,
-            month: None,
-            month_code: None,
-            era: None,
-            era_year: None,
-        }
-    }
-
-    pub const fn with_era(mut self, era: Option<TinyAsciiStr<19>>) -> Self {
-        self.era = era;
-        self
-    }
-
-    pub const fn with_era_year(mut self, era_year: Option<i32>) -> Self {
-        self.era_year = era_year;
-        self
-    }
-
-    pub const fn with_year(mut self, year: i32) -> Self {
-        self.year = Some(year);
-        self
-    }
-
-    pub const fn with_optional_year(mut self, year: Option<i32>) -> Self {
-        self.year = year;
-        self
-    }
-
-    pub const fn with_month(mut self, month: u8) -> Self {
-        self.month = Some(month);
-        self
-    }
-
-    pub const fn with_optional_month(mut self, month: Option<u8>) -> Self {
-        self.month = month;
-        self
-    }
-
-    pub const fn with_month_code(mut self, month_code: MonthCode) -> Self {
-        self.month_code = Some(month_code);
-        self
-    }
-
-    pub const fn with_optional_month_code(mut self, month_code: Option<MonthCode>) -> Self {
-        self.month_code = month_code;
-        self
-    }
-}
-
-impl YearMonthCalendarFields {
-    pub(crate) fn try_from_year_month(year_month: &PlainYearMonth) -> TemporalResult<Self> {
-        let (year, era, era_year) = if year_month.era().is_some() {
-            (
-                None,
-                year_month
-                    .era()
-                    .map(|t| TinyAsciiStr::<19>::try_from_utf8(t.as_bytes()))
-                    .transpose()
-                    .map_err(|e| TemporalError::general(format!("{e}")))?,
-                year_month.era_year(),
-            )
-        } else {
-            (Some(year_month.year()), None, None)
-        };
-        Ok(Self {
-            year,
-            month: Some(year_month.month()),
-            month_code: Some(year_month.month_code()),
-            era,
-            era_year,
-        })
-    }
-
-    pub fn is_empty(&self) -> bool {
-        *self == Self::new()
-    }
-
-    crate::impl_with_fallback_method!(with_fallback_year_month, CalendarFields, () PlainYearMonth);
-    crate::impl_field_keys_to_ignore!(());
-}
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum ResolutionType {
