+++ src/builtins/core/date.rs
@@ -41,123 +41,6 @@ pub struct PartialDate {
     pub calendar: Calendar,
 }
 
-/// The return value of CalendarFieldKeysToIgnore
-#[derive(Copy, Clone, Default)]
-pub(crate) struct FieldKeysToIgnore {
-    /// Ignore all era fields (era-year, era)
-    pub era: bool,
-    /// Ignore arithmetical year
-    pub arithmetical_year: bool,
-    /// Ignore all month fields (month, month-code)
-    pub month: bool,
-}
-
-// Use macro to impl fallback methods to avoid having a trait method.
-#[doc(hidden)]
-#[macro_export]
-macro_rules! impl_with_fallback_method {
-    ($method_name:ident, $fields_type:ident, ( $(with_day: $day:ident)? ) $component_type:ty) => {
-        pub(crate) fn $method_name(&self, fallback: &$component_type, calendar: icu_calendar::AnyCalendarKind, overflow: ArithmeticOverflow) -> TemporalResult<Self> {
-            let keys_to_ignore = self.field_keys_to_ignore(calendar);
-            let mut era = self.era;
-
-            let mut era_year = self.era_year;
-            let mut year = self.year;
-
-            if !keys_to_ignore.era {
-                if era.is_none() {
-                    era =
-                        fallback.era().map(|e| {
-                            TinyAsciiStr::<19>::try_from_utf8(e.as_bytes())
-                                .map_err(|_| TemporalError::assert().with_message("Produced invalid era code"))
-                        })
-                        .transpose()?
-                }
-                if era_year.is_none() {
-                    era_year = fallback.era_year();
-                }
-            }
-            if !keys_to_ignore.arithmetical_year {
-                if year.is_none() {
-                    year = Some(fallback.year());
-                }
-            }
-
-            let (month, month_code) = match (self.month, self.month_code) {
-                (Some(month), Some(mc)) => (Some(month), Some(mc)),
-                (Some(month), None) => {
-                    let month_maybe_clamped = if overflow == ArithmeticOverflow::Constrain {
-                        // TODO (manishearth) this should be managed by ICU4X
-                        // https://github.com/unicode-org/icu4x/issues/6790
-                        month.clamp(1, 12)
-                    } else {
-                        month
-                    };
-
-                    (Some(month_maybe_clamped), Some(month_to_month_code(month_maybe_clamped)?))
-                }
-                (None, Some(mc)) => (Some(mc.to_month_integer()).map(Into::into), Some(mc)),
-                (None, None) if !keys_to_ignore.month => (
-                    Some(fallback.month()).map(Into::into),
-                    Some(fallback.month_code()),
-                ),
-                // This should currently be unreachable, but it may change as CalendarFieldKeysToIgnore
-                // changes
-                (None, None) => (None, None)
-            };
-            #[allow(clippy::needless_update)] {
-                Ok(Self {
-                    year,
-                    month,
-                    month_code,
-                    $($day: Some(self.day.unwrap_or(fallback.day().into())),)?
-                    era,
-                    era_year,
-                })
-            }
-        }
-    };
-}
-#[doc(hidden)]
-#[macro_export]
-macro_rules! impl_field_keys_to_ignore {
-    (( $(with_day: $day:ident)? )) => {
-        /// <https://tc39.es/proposal-temporal/#sec-temporal-calendarfieldkeystoignore>
-        fn field_keys_to_ignore(&self, calendar: icu_calendar::AnyCalendarKind) -> $crate::builtins::core::date::FieldKeysToIgnore {
-            let mut keys = $crate::builtins::core::date::FieldKeysToIgnore::default();
-            // All calendars have months/month codes
-            if self.month.is_some() || self.month_code.is_some() {
-                keys.month = true;
-            }
-            if Calendar::calendar_has_eras(calendar) {
-                // We should clear years only if the calendar has eras
-                if self.year.is_some() || self.era_year.is_some() || self.era.is_some() {
-                    keys.era = true;
-                    keys.arithmetical_year = true;
-                }
-
-                // In a calendar such as "japanese" where eras do not start and end at year and/or month boundaries, note that the returned
-                // List should contain era and era-year if keys contains day, month, or month-code
-                // (not only if it contains era, era-year, or year, as in the example above) because it's possible for
-                // changing the day or month to cause a conflict with the era.
-                if calendar == icu_calendar::AnyCalendarKind::Japanese {
-                    if self.month.is_some() || self.month_code.is_some() {
-                        keys.era = true;
-                    }
-
-                    $(
-                        if self.$day.is_some() {
-                            keys.era = true;
-                        }
-                    )?
-                }
-            }
-
-            keys
-        }
-    };
-}
-
 /// Convenience methods for building a `PartialDate`
 impl PartialDate {
     pub const fn new() -> Self {
